refactor: moved quest components into its folder
