+++ packages/web/components/Quest/Skills.tsx
